Fix IE form validation bug
Supposing that the select has the "required" attribute; if we submit a
form, IE correctly validates the form and warns us. But selecting
something later do not clear this validation error. Assigning the
selectedIndex property of select does not work, either. The only thing
that worked was to temporarily remove and then re-insert the "required"
attribute.

Dear IE, please burn in hell.